swc
Struct Compiler

pub struct Compiler {
    pub cm: Arc<SourceMap>,
    /* private fields */
}

All methods accept Handler, which is a storage for errors.

The caller should check if the handler contains any errors after calling method.

Fields
cm: Arc<SourceMap>

CodeMap

Implementations

impl Compiler

These are low-level apis.

pub fn comments(&self) -> &SwcComments

pub fn run<R, F>(&self, op: F) -> R
where F: FnOnce() -> R,

Runs op in current compiler’s context.

Note: Other methods of Compiler already uses this internally.

pub fn parse_js( &self, fm: Arc<SourceFile>, handler: &Handler, target: EsVersion, syntax: Syntax, is_module: IsModule, comments: Option<&dyn Comments>, ) -> Result<Program, Error>

This method parses a javascript / typescript file

pub fn print<T>( &self, node: &T, source_file_name: Option<&str>, output_path: Option<PathBuf>, inline_sources_content: bool, target: EsVersion, source_map: SourceMapsConfig, source_map_names: &AHashMap<BytePos, JsWord>, orig: Option<&SourceMap>, minify: bool, comments: Option<&dyn Comments>, emit_source_map_columns: bool, ascii_only: bool, ) -> Result<TransformOutput, Error>
where T: Node + VisitWith<IdentCollector>,

Converts ast node to source string and sourcemap.

This method receives target file path, but does not write file to the path. See: https://github.com/swc-project/swc/issues/1255

impl Compiler

High-level apis.

pub fn new(cm: Arc<SourceMap>) -> Self

pub fn read_config( &self, opts: &Options, name: &FileName, ) -> Result<Option<Config>, Error>

pub fn parse_js_as_input<'a, P>( &'a self, fm: Lrc<SourceFile>, program: Option<Program>, handler: &'a Handler, opts: &Options, name: &FileName, comments: Option<&'a SingleThreadedComments>, before_pass: impl 'a + FnOnce(&Program) -> P, ) -> Result<Option<BuiltInput<impl 'a + Fold>>, Error>
where P: 'a + Fold,

This method returns None if a file should be skipped.

This method handles merging of config.

This method does not parse module.

pub fn run_transform<F, Ret>( &self, handler: &Handler, external_helpers: bool, op: F, ) -> Ret
where F: FnOnce() -> Ret,

pub fn transform( &self, handler: &Handler, program: Program, external_helpers: bool, pass: impl Fold, ) -> Program

pub fn process_js_with_custom_pass<P1, P2>( &self, fm: Arc<SourceFile>, program: Option<Program>, handler: &Handler, opts: &Options, comments: SingleThreadedComments, custom_before_pass: impl FnOnce(&Program) -> P1, custom_after_pass: impl FnOnce(&Program) -> P2, ) -> Result<TransformOutput, Error>
where P1: Fold, P2: Fold,

custom_after_pass is applied after swc transforms are applied.

program: If you already parsed Program, you can pass it.

Guarantee

swc invokes custom_before_pass after

Handling decorators, if configured

Applying resolver

Stripping typescript nodes

This means, you can use noop_visit_type, noop_fold_type and noop_visit_mut_type in your visitor to reduce the binary size.

pub fn process_js_file( &self, fm: Arc<SourceFile>, handler: &Handler, opts: &Options, ) -> Result<TransformOutput, Error>

pub fn minify( &self, fm: Arc<SourceFile>, handler: &Handler, opts: &JsMinifyOptions, ) -> Result<TransformOutput, Error>

pub fn process_js( &self, handler: &Handler, program: Program, opts: &Options, ) -> Result<TransformOutput, Error>

You can use custom pass with this method.

There exists a PassBuilder to help building custom passes.