Struct swc::PassBuilder

pub struct PassBuilder<'a, 'b, P: Fold> { /* fields omitted */ }

Builder is used to create a high performance Compiler.

Implementations

impl<'a, 'b, P: Fold> PassBuilder<'a, 'b, P>

pub fn new(
cm: &'a Arc<SourceMap>,
handler: &'b Handler,
loose: bool,
global_mark: Mark,
pass: P
) -> Self

pub fn then<N>(self, next: N) -> PassBuilder<'a, 'b, AndThen<P, N>> where
N: Fold,

pub fn skip_helper_injection(self, skip: bool) -> Self

pub fn fixer(self, enable: bool) -> Self

Note: fixer is enabled by default.

pub fn hygiene(self, enable: bool) -> Self

Note: hygiene is enabled by default.

pub fn const_modules(
self,
globals: HashMap<JsWord, HashMap<JsWord, String>>
) -> PassBuilder<'a, 'b, impl Fold>

pub fn inline_globals(
self,
c: GlobalPassOption
) -> PassBuilder<'a, 'b, impl Fold>

pub fn target(self, target: JscTarget) -> Self

pub fn preset_env(self, env: Option<Config>) -> Self

pub fn finalize<'cmt>(
self,
syntax: Syntax,
module: Option<ModuleConfig>,
comments: Option<&'cmt dyn Comments>
) -> impl 'cmt + Fold where
P: 'cmt,

Arguments

module

Use None if you want swc to emit import statements.

Returned pass includes

compatibility helper

module handler

helper injector

identifier hygiene handler if enabled

fixer if enabled