Struct swc::PassBuilder

pub struct PassBuilder<'a, 'b, P: Pass> { /* fields omitted */ }

Builder is used to create a high performance Compiler.

Implementations

impl<'a, 'b, P: Pass> PassBuilder<'a, 'b, P>

pub fn new(
cm: &'a Arc<SourceMap>,
handler: &'b Handler,
loose: bool,
global_mark: Mark,
pass: P
) -> Self

pub fn then<N>(self, next: N) -> PassBuilder<'a, 'b, AndThen<P, N>>

pub fn fixer(self, enable: bool) -> Self

Note: fixer is enabled by default.

pub fn hygiene(self, enable: bool) -> Self

Note: hygiene is enabled by default.

pub fn const_modules(
self,
globals: HashMap<JsWord, HashMap<JsWord, String>>
) -> PassBuilder<'a, 'b, impl Pass>

pub fn inline_globals(
self,
c: GlobalPassOption
) -> PassBuilder<'a, 'b, impl Pass>

pub fn strip_typescript(self) -> PassBuilder<'a, 'b, impl Pass>

pub fn target(self, target: JscTarget) -> Self

pub fn preset_env(self, env: Option<Config>) -> Self

pub fn finalize(
self,
root_mark: Mark,
syntax: Syntax,
module: Option<ModuleConfig>
) -> impl Pass

Arguments

module

Use None if you want swc to emit import statements.

Returned pass includes

compatibility helper

module handler

helper injector

identifier hygiene handler if enabled

fixer if enabled